Added authentication and authorisation

+++ src/main/java/com/deepak/projects/airBnbApp/service/HotelServiceImpl.java
@@ -5,13 +5,16 @@ import com.deepak.projects.airBnbApp.dto.HotelInfoDto;
 import com.deepak.projects.airBnbApp.dto.RoomDto;
 import com.deepak.projects.airBnbApp.entity.Hotel;
 import com.deepak.projects.airBnbApp.entity.Room;
+import com.deepak.projects.airBnbApp.entity.User;
 import com.deepak.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.deepak.projects.airBnbApp.exception.UnAuthorisedException;
 import com.deepak.projects.airBnbApp.repository.HotelRepository;
 import com.deepak.projects.airBnbApp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,6 +34,10 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating new hotel with name: {}", hotelDto.getName());
         Hotel hotel= modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user= (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel= hotelRepository.save(hotel);
         log.info("Created a new hotel with id: {}", hotelDto.getId());
         return modelMapper.map(hotel,HotelDto.class);
@@ -39,10 +46,15 @@ public class HotelServiceImpl implements HotelService{
     @Override
     public HotelDto getHotelById(Long id) {
         log.info("Getting hotel with id: {}", id);
+
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id: "+id));
 
+//        User user= (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+//        if(!user.equals(hotel.getOwner())){
+//            throw new UnAuthorisedException("This user doesnt not own this hotel with id: "+id);
+//        }
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -52,6 +64,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id: "+id));
+
+        User user= (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user doesn not own this hotel with id: "+id);
+        }
+
         modelMapper.map(hotelDto,hotel);
         hotel.setId(id);
         hotel= hotelRepository.save(hotel);
@@ -66,6 +84,10 @@ public class HotelServiceImpl implements HotelService{
                 .findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id: "+id));
 
+        User user= (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user doesn not own this hotel with id: "+id);
+        }
 
         for(Room room: hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
@@ -82,6 +104,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel= hotelRepository
                 .findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id: "+id));
+
+        User user= (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user doesn not own this hotel with id: "+id);
+        }
+
         hotel.setActive(true);
         //Create inventories for this hotel for all rooms(assuming only do it once)
         for(Room room: hotel.getRooms()){
@@ -89,6 +117,7 @@ public class HotelServiceImpl implements HotelService{
         }
     }
 
+    //public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel= hotelRepository
